+++ ModelLibrary/Utility/NormalDisribution.cs
@@ -11,20 +11,20 @@ namespace Utility
     {
         TrapezoidApproximation integral = new TrapezoidApproximation();
 
-        public decimal normalDistribution(decimal x)
+        public double normalDistribution(double x)
         {
-            Func<decimal, decimal> f = e =>
+            Func<double, double> f = e =>
             {
                 return normalDistProbabilityDensity(e);
             };
 
             const int partition = 10000000;
-            decimal result = integral.integrate(-100000, x, partition, f);
+            double result = integral.integrate(-100000, x, partition, f);
 
             return result;
         }
 
-        private decimal normalDistProbabilityDensity(decimal x)
+        private double normalDistProbabilityDensity(double x)
         {
             return Math.Exp(- 0.5 * x * x) / Math.Sqrt(2.0 * Math.PI);
